move all mcref code to src/mcref dir
