+++ src/mcref/HypothesisPrinter.h
